fix(test) Make dir before calling click app The app checks if the folder exists
